+++ src/main/java/softuni/exam/repository/PlayerRepository.java
@@ -4,7 +4,8 @@ package softuni.exam.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import softuni.exam.domain.entitys.Player;
+
 @Repository
-public interface PlayerRepository extends JpaRepository<Player,Integer> {
+public interface PlayerRepository extends JpaRepository<Player,Long> {
 
 }
