+++ src/main/java/softuni/exam/repository/TeamRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import softuni.exam.domain.entitys.Team;
 
 @Repository
-public interface TeamRepository  extends JpaRepository<Team,Integer> {
+public interface TeamRepository  extends JpaRepository<Team,Long> {
 
 }
